+++ foundation/blockchain/genesis/genesis.go
@@ -9,13 +9,13 @@ import (
 
 // Genesis represents the genesis file.
 type Genesis struct {
-	Date          time.Time       `json:"date"`
-	ChainID       uint            `json:"chain_id"`        // The chain id represents an unique id for this running instance.
-	TransPerBlock uint            `json:"trans_per_block"` // The maximum number of transactions that can be in a block.
-	Difficulty    uint            `json:"difficulty"`      // How difficult it needs to be to solve the work problem.
-	MiningReward  uint            `json:"mining_reward"`   // Reward for mining a block.
-	GasPrice      uint            `json:"gas_price"`       // Fee paid for each transaction mined into a block.
-	Balances      map[string]uint `json:"balances"`
+	Date          time.Time         `json:"date"`
+	ChainID       uint16            `json:"chain_id"`        // The chain id represents an unique id for this running instance.
+	TransPerBlock uint16            `json:"trans_per_block"` // The maximum number of transactions that can be in a block.
+	Difficulty    uint16            `json:"difficulty"`      // How difficult it needs to be to solve the work problem.
+	MiningReward  uint64            `json:"mining_reward"`   // Reward for mining a block.
+	GasPrice      uint64            `json:"gas_price"`       // Fee paid for each transaction mined into a block.
+	Balances      map[string]uint64 `json:"balances"`
 }
 
 // =============================================================================
